Version6 - No Internet Connection Added when user not connected to internet in Dashboard_Activity.java - Shimmer Effect added in all screens - All Data Store via User_id.

+++ app/src/main/java/com/example/orderfruit/Interface/InterfaceData.java
@@ -8,7 +8,8 @@ public interface InterfaceData {
 
     void onNothingSelected(AdapterView<?> adapterView);
 
-    void Favourite_fruite(int id, int val);
+    void Favourite_fruite(long id, int val);
+
 
 
 }
